Refine ice and ground movement logic. Bug fixes. Add a bit of level design to level 7
Improved player movement handling on ice and ground by adjusting conditionals and removing debug prints. Fixed many bugs related to movement on ice physics and logic, such as fixed dashing in midair if the player was standing on ice, fixed a bug in dashing direction. Furthermore, allowed the player to move in the air with acceleration and deceleration if the player jumped from ice. Updated level 7 scene, adding a bit more of level design.